Fix "Design Rule Checker moves elements and traces off grid"
Patch from Ineiev <ineiev@gmail.com> to keep the grid from changing
just because DRC ran.

+++ src/misc.c
@@ -599,7 +599,7 @@ GetDataBoundingBox (DataTypePtr Data)
 void
 CenterDisplay (LocationType X, LocationType Y, Boolean Delta)
 {
-  int save_grid = PCB->Grid;
+  double save_grid = PCB->Grid;
   PCB->Grid = 1;
   if (Delta)
     {
